Qt6: remove useless casts
git-svn-id: svn://scribus.net/trunk/Scribus@24903 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/navigator.cpp
@@ -55,21 +55,21 @@ Navigator::Navigator(QWidget *parent, int size, int pageNr, ScribusView* view, c
 void Navigator::mouseMoveEvent(QMouseEvent *m)
 {
 	drawMark(m->position().x(), m->position().y());
-	emit Coords(static_cast<double>(m->position().x()) / static_cast<double>(pmx.width()),
-				static_cast<double>(m->position().y()) / static_cast<double>(pmx.height()));
+	emit Coords(m->position().x() / static_cast<double>(pmx.width()),
+				m->position().y() / static_cast<double>(pmx.height()));
 }
 
 void Navigator::mousePressEvent(QMouseEvent *m)
 {
 	drawMark(m->position().x(), m->position().y());
-	emit Coords(static_cast<double>(m->position().x()) / static_cast<double>(pmx.width()),
-				static_cast<double>(m->position().y()) / static_cast<double>(pmx.height()));
+	emit Coords(m->position().x() / static_cast<double>(pmx.width()),
+				m->position().y() / static_cast<double>(pmx.height()));
 }
 
 void Navigator::mouseReleaseEvent(QMouseEvent *m)
 {
-	emit Coords(static_cast<double>(m->position().x()) / static_cast<double>(pmx.width()),
-				static_cast<double>(m->position().y()) / static_cast<double>(pmx.height()));
+	emit Coords(m->position().x() / static_cast<double>(pmx.width()),
+				m->position().y() / static_cast<double>(pmx.height()));
 }
 
 void Navigator::paintEvent(QPaintEvent *e)
